# back/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/organizers": {
            "get": {
                "description": "Get all organizers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organizers"
                ],
                "summary": "Get all organizers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/easynight_internal_models.Organizer"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Register as organizer",
                "parameters": [
                    {
                        "description": "Registration payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.RegistrationPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/organizers/id/{id}": {
            "get": {
                "description": "Get an organizer by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organizers"
                ],
                "summary": "Get an organizer by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organizer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/easynight_internal_models.Organizer"
                        }
                    }
                }
            }
        },
        "/organizers/{id}": {
            "get": {
                "description": "Get an organizer by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organizers"
                ],
                "summary": "Get an organizer by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organizer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/easynight_internal_models.Organizer"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an organizer by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organizers"
                ],
                "summary": "Update an organizer by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organizer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Organizer Input",
                        "name": "organizer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.OrganizerInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/easynight_internal_models.Organizer"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an organizer by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organizers"
                ],
                "summary": "Delete an organizer by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organizer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "easynight_internal_models.Admin": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/easynight_internal_models.User"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "easynight_internal_models.Customer": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "reservations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/easynight_internal_models.Reservation"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/easynight_internal_models.User"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "easynight_internal_models.Event": {
            "type": "object",
            "properties": {
                "banner": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "location": {
                    "type": "string"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/easynight_internal_models.Message"
                    }
                },
                "organizers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/easynight_internal_models.Organizer"
                    }
                },
                "participantNumber": {
                    "type": "integer"
                },
                "place": {
                    "type": "string"
                },
                "rates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/easynight_internal_models.Rate"
                    }
                },
                "reservations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/easynight_internal_models.Reservation"
                    }
                },
                "tag": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "easynight_internal_models.Message": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "event": {
                    "$ref": "#/definitions/easynight_internal_models.Event"
                },
                "eventID": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "organizer": {
                    "$ref": "#/definitions/easynight_internal_models.Organizer"
                },
                "organizerID": {
                    "type": "string"
                },
                "sender": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "easynight_internal_models.Notification": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "string"
                },
                "isOpened": {
                    "type": "boolean"
                },
                "receiver": {
                    "$ref": "#/definitions/easynight_internal_models.User"
                },
                "receiverID": {
                    "type": "string"
                },
                "sender": {
                    "$ref": "#/definitions/easynight_internal_models.User"
                },
                "senderID": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "easynight_internal_models.Organizer": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "firstname": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/easynight_internal_models.Message"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/easynight_internal_models.User"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "easynight_internal_models.Rate": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "event": {
                    "$ref": "#/definitions/easynight_internal_models.Event"
                },
                "eventID": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "note": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/easynight_internal_models.User"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "easynight_internal_models.Reservation": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "customer": {
                    "$ref": "#/definitions/easynight_internal_models.Customer"
                },
                "customerID": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "event": {
                    "$ref": "#/definitions/easynight_internal_models.Event"
                },
                "eventID": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "qrcode": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "easynight_internal_models.User": {
            "type": "object",
            "properties": {
                "admins": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/easynight_internal_models.Admin"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "customers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/easynight_internal_models.Customer"
                    }
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "notificationReceived": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/easynight_internal_models.Notification"
                    }
                },
                "notificationSent": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/easynight_internal_models.Notification"
                    }
                },
                "organizers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/easynight_internal_models.Organizer"
                    }
                },
                "password": {
                    "type": "string"
                },
                "rates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/easynight_internal_models.Rate"
                    }
                },
                "role": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "internal_controllers.OrganizerInput": {
            "type": "object",
            "properties": {
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_controllers.RegistrationPayload": {
            "type": "object",
            "required": [
                "email",
                "firstname",
                "lastname"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
    }
}